% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adif_headers.R
\name{get_adif_headers}
\alias{get_adif_headers}
\title{Read adif header lines until an opening "<" is found}
\usage{
get_adif_headers(conn)
}
\arguments{
\item{conn}{A file connection}
}
\value{
A character vector that is the line that caused the exit condition.
}
\description{
This needs to be changes to character base but will probably work
for the skcc logger for now
}
\examples{

src <- file.path(getwd(),"adif","tests","sample.adi")
conn <- file(src, open="r")
res <- get_adif_headers(conn)
close(conn)
res
}
