% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adif_record.R
\name{get_adif_record}
\alias{get_adif_record}
\title{Read adif record lines until an  "<eor>" is found}
\usage{
get_adif_record(conn)
}
\arguments{
\item{conn}{A file connection}
}
\value{
a character vector:
    If character(1) then it is the line that caused the exit condition.
    If character(0) then it is the end of file
}
\description{
This needs to be changed to character base but will probably work
for the skcc logger for now
}
\examples{
src <- file.path(getwd(),"adif","tests","sample.adi")
conn <- file(src, open="r")
get_adif_headers(conn)
res <- get_adif_record(conn)
close(conn)
res
}
\keyword{adif,}
\keyword{headers}
