% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isEOR}
\alias{isEOR}
\title{Check adif lines for <eor>}
\usage{
isEOR(line)
}
\arguments{
\item{line}{a line from an adif file}
}
\value{
a boolean:
}
\description{
This needs to be changes to character base but will probably work
for the skcc logger for now
}
\keyword{adif,}
\keyword{tag}
