% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeaders.R
\name{getHeaders}
\alias{getHeaders}
\title{Read adif header lines until an opening "<" is found}
\usage{
getHeaders(conn)
}
\description{
This needs to be changes to character base but will probably work
 for the skcc logger for now
}
\details{
@param conn A file connection
 @export
 @return A character vector that is the line that caused the exit condition.
 @examples
 src <- file.path(getwd(),"inst","tests","sample.adi")
 conn <- file(src, open="r")
 res <- getHeaders(conn)
 close(conn)
 res
}
