% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeaders.R
\name{getHeaders}
\alias{getHeaders}
\title{Read adif header lines until an opening "<" is found}
\usage{
getHeaders(con)
}
\description{
This needs to be changes to character base but will probably work
 for the skcc logger for now
}
\details{
@param con A file connection
 @examples
 getHeaders(file("test.adi", "r"))
 @return character vector:
 If character(1) then it is the line that caused the exit condition.
 If character(0) then it is the end of file
}
