% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecord.R
\name{getRecord}
\alias{getRecord}
\title{Read adif record lines until an  "<eor>" is found}
\usage{
getRecord(con)
}
\description{
This needs to be changed to character base but will probably work
 for the skcc logger for now
}
\details{
@param con A file connection
 @keywords adif, headers
 @return a character vector:
     If character(1) then it is the line that caused the exit condition.
     If character(0) then it is the end of file
 @export
 @examples
     con = file(fPath, "r")
     currentLine <- getRecord(con)
     while ( length(currentLine > 0 ) )  {
        currentLine <- getRecord(con)
        }
}
