% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isEOH}
\alias{isEOH}
\title{Check adif lines for <eoh>}
\usage{
isEOH(line)
}
\description{
This needs to be changes to character base but will probably work
 for the skcc logger for now
}
\details{
@param line
 @keywords adif, tag
 @return a boolean:
 @export
}
