% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isEOR}
\alias{isEOR}
\title{Check adif lines for <eor>}
\usage{
isEOR(line)
}
\description{
This needs to be changes to character base but will probably work
 for the skcc logger for now
 @param line
 @keywords adif, tag
 @return a boolean:
 @export
}
